Squashed commit of the following:
      - portable_optimized implementations of fully_connected, conv and depthwise-conv
      - test suite adapted to run correctly on (at last x86) 32-bit targets
      - Micro-allocator now supports mixing Scratch and Persistent Buffer allocation in prepare phase
      - Minor coverage/robustness improvements in test suite.
      - MinGW target support "windows_x86_makefile.inc"

+++ tensorflow/lite/micro/kernels/l2norm.cc
@@ -51,7 +51,7 @@ TfLiteStatus Prepare(TfLiteContext* context, TfLiteNode* node) {
   TF_LITE_ENSURE_TYPES_EQ(context, input->type, output->type);
 
   if (output->type == kTfLiteUInt8 || output->type == kTfLiteInt8) {
-    TF_LITE_ENSURE_EQ(context, output->params.scale, (1. / 128.));
+    TF_LITE_ENSURE_EQ(context, (double)output->params.scale, (1. / 128.));
     if (output->type == kTfLiteUInt8) {
       TF_LITE_ENSURE_EQ(context, output->params.zero_point, 128);
     }
